requirements.txt
asgiref==3.8.1
Django==5.1
django-ckeditor==6.7.1
django-ckeditor-5==0.2.13
django-jazzmin==3.0.0
django-js-asset==2.2.0
djangorestframework==3.15.2
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
pillow==10.4.0
pytz==2024.1
PyYAML==6.0.2
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
